Switch to using a fork of the 'gltf-rs/gltf' crate.
We're going to need to use a forked version of these crates for now.

For one thing, we need to implement FB_material_variants, one way or
another. It could be added to the gltf-rs/gltf fork, as one of the
hard-coded extensions, but that seems irritating. It's really only a
concern for glTFVariantMeld.

The other is to implement some simple way to implement user extension
support; something like https://github.com/gltf-rs/gltf/issues/231.